% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_cols}
\alias{.assert_cols}
\title{Assert that a data.frame has all required columns}
\usage{
.assert_cols(df, cols, label = deparse(substitute(df)))
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{Character vector of required column names.}

\item{label}{Informative label for error messages (e.g. \code{"col_info"}).}
}
\description{
Assert that a data.frame has all required columns
}
\keyword{internal}
