% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.brace_split}
\alias{.brace_split}
\title{Split a \verb{\{A\}\{B\}\{C\}} brace-string; thin convenience wrapper used in loops
where NA should silently yield \code{character(0)}.}
\usage{
.brace_split(s)
}
\description{
Split a \verb{\{A\}\{B\}\{C\}} brace-string; thin convenience wrapper used in loops
where NA should silently yield \code{character(0)}.
}
\keyword{internal}
