% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.non_na_rows}
\alias{.non_na_rows}
\title{Return rows of a data.frame where a column is non-NA}
\usage{
.non_na_rows(df, col)
}
\description{
Return rows of a data.frame where a column is non-NA
}
\keyword{internal}
