% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dsl.R
\name{.parse_braces_vec}
\alias{.parse_braces_vec}
\title{Vectorised wrapper: parse a character vector of brace-list strings}
\usage{
.parse_braces_vec(sv, allow_bare = FALSE)
}
\arguments{
\item{sv}{Character vector (NAs allowed).}

\item{allow_bare}{Passed to \code{\link[=.parse_braces]{.parse_braces()}}.}
}
\value{
A list of character vectors, same length as \code{sv}.
}
\description{
Vectorised wrapper: parse a character vector of brace-list strings
}
\keyword{internal}
