% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_cutoff.R
\name{md_apply_cutoff}
\alias{md_apply_cutoff}
\title{Apply threshold-based scoring to create discretised columns}
\usage{
md_apply_cutoff(data, cutoff_rows)
}
\arguments{
\item{data}{data.frame  --  the clinical data}

\item{cutoff_rows}{data.frame  --  the \verb{$cutoff} sub-table from
md_extract_var_roles(); must have columns
\code{col_name}, \code{transform}, \code{cutoff}, \code{var.score}}
}
\value{
data with new numeric columns added; rows never added/removed
}
\description{
Apply threshold-based scoring to create discretised columns
}
