% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_strata.R
\name{md_apply_strata}
\alias{md_apply_strata}
\title{Filter data rows to allowed strata values}
\usage{
md_apply_strata(data, strata_rows)
}
\arguments{
\item{data}{data.frame  --  the clinical data}

\item{strata_rows}{data.frame  --  the \verb{$strata_in} sub-table from
md_extract_var_roles(); must have columns
\code{col_name} and \code{strata_in}}
}
\value{
data with non-matching rows removed; columns are unchanged
}
\description{
Filter data rows to allowed strata values
}
