% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{mf.bmi}
\alias{mf.bmi}
\title{Compute BMI from weight (kg) and height (cm) columns}
\usage{
mf.bmi(data, x1, x2, ..., id_cols = character(0))
}
\arguments{
\item{data}{data.frame}

\item{x1}{column name -- weight in kg}

\item{x2}{column name -- height in cm}

\item{...}{ignored}

\item{id_cols}{character vector of ID column names}
}
\value{
numeric vector
}
\description{
Compute BMI from weight (kg) and height (cm) columns
}
