% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{mf.days_to_year}
\alias{mf.days_to_year}
\title{Convert a numeric days column to whole years (floor division)}
\usage{
mf.days_to_year(data, x1, ..., id_cols = character(0))
}
\arguments{
\item{data}{data.frame}

\item{x1}{column name -- numeric days value}

\item{...}{ignored}

\item{id_cols}{character vector of ID column names}
}
\value{
integer vector (floor of days / 365)
}
\description{
Convert a numeric days column to whole years (floor division)
}
