% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{mf.text_parse}
\alias{mf.text_parse}
\title{Evaluate an R expression string with data columns in scope}
\usage{
mf.text_parse(data, x1, ..., id_cols = character(0))
}
\arguments{
\item{data}{data.frame}

\item{x1}{character(1) -- an R expression string to evaluate}

\item{...}{ignored}

\item{id_cols}{character vector of ID column names}
}
\value{
result of evaluating the expression (typically a vector)
}
\description{
Evaluates \code{x1} as an R expression using \code{with(data, ...)},
making all columns of \code{data} available by name.
}
