% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_functions.R
\name{mf.wday}
\alias{mf.wday}
\title{Extract day of week from a date column}
\usage{
mf.wday(data, x, ..., id_cols = character(0))
}
\arguments{
\item{data}{data.frame}

\item{x}{column name -- a date column}

\item{...}{ignored}

\item{id_cols}{character vector of ID column names}
}
\value{
integer vector (1 = Sunday, 7 = Saturday)
}
\description{
Returns an integer (1 = Sunday, 7 = Saturday) via
\code{lubridate::wday()}.
}
